% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_raw}
\alias{as_raw}
\title{As Raw}
\usage{
as_raw(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
(any)
}
\description{
Convert a value (usually a tensor from some framework) to a \code{raw()} vector.
}
