% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dtype}
\alias{dtype}
\title{Data Type}
\usage{
dtype(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
(any)
}
\description{
Get the data type of a tensor.
The return type depends on the specific tensor implementation.
}
