% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bib.R
\name{format_bib}
\alias{format_bib}
\alias{cite_bib}
\title{Format Bibentries in Roxygen}
\usage{
format_bib(..., bibentries = NULL)

cite_bib(..., bibentries = NULL)
}
\arguments{
\item{...}{(\code{character()})\cr
One or more names of bibentries.}

\item{bibentries}{(named \code{list()})\cr
Named list of bibentries. If \code{NULL}, looks up an object called
\code{bibentries} in the calling environment.}
}
\value{
(\code{character(1)}).
}
\description{
Operates on a named list of \code{\link[=bibentry]{bibentry()}} entries and formats them for
documentation with \CRANpkg{roxygen2}.
\itemize{
\item \code{format_bib()} is intended to be called in the \verb{@references} section and
formats the complete entry using \code{\link[tools:toRd]{tools::toRd()}}.
\item \code{cite_bib()} returns a short inline citation in the format
\code{"LastName (Year)"}.
}
}
\examples{
bibentries = list(R = citation())
format_bib("R")
cite_bib("R")
}
