% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dims.R
\name{get_dims}
\alias{get_dims}
\title{Get the dimensions of a data object}
\usage{
get_dims(data)
}
\arguments{
\item{data}{(\code{any})\cr
The data object to get the dimensions of.}
}
\value{
(\code{integer()})
}
\description{
Get the "dimension" of an R array or vector.
}
