% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of.R
\name{list_of}
\alias{list_of}
\title{Create a list property}
\usage{
list_of(class, ...)
}
\arguments{
\item{class}{The class of the list.}

\item{...}{The properties of the list.}
}
\description{
Create a list property.
}
