% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lru_cache.R
\name{LRUCache}
\alias{LRUCache}
\title{Least Recently Used Cache}
\description{
An implementation of a least-recently-used cache built on top of \code{\link[utils:hashtab]{utils::hashtab}}.
Therefore, arbitrary keys can be used, as opposed to the implementation in \CRANpkg{cachemem},
which relies on environments.
}
\details{
This LRU cache is implemented as a combination of a hashmap and a doubly linked list.
The hashmap is used for lookups and the doubly linked list is used to maintain the
ordering of most recently used to least recently used items.
Whenever an element is added or accessed, it is moved to the front of the list.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{capacity}}{(\code{integer(1)})\cr
The maximum capacity of the cache.}

    \item{\code{size}}{(\code{integer(1)})\cr
The number of items currently stored.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LRUCache-initialize}{\code{LRUCache$new()}}
    \item \href{#method-LRUCache-get}{\code{LRUCache$get()}}
    \item \href{#method-LRUCache-set}{\code{LRUCache$set()}}
    \item \href{#method-LRUCache-has}{\code{LRUCache$has()}}
    \item \href{#method-LRUCache-remove}{\code{LRUCache$remove()}}
    \item \href{#method-LRUCache-clear}{\code{LRUCache$clear()}}
    \item \href{#method-LRUCache-keys_mru_to_lru}{\code{LRUCache$keys_mru_to_lru()}}
    \item \href{#method-LRUCache-clone}{\code{LRUCache$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-initialize}{}}}
\subsection{\code{LRUCache$new()}}{
  Initialize a new LRU cache with a given \code{capacity}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$new(capacity)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{capacity}}{(\code{integer(1)})\cr
Number of items the cache holds.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-get"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-get}{}}}
\subsection{\code{LRUCache$get()}}{
  Get the value for \code{key} and mark it as most-recently-used.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$get(key, default = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{any})\cr
Key.}
      \item{\code{default}}{(\code{any})\cr
Default value returned by \code{get()} when \code{key} is not present.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-set"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-set}{}}}
\subsection{\code{LRUCache$set()}}{
  Set the value for \code{key}, updating recency and evicting LRU as needed.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$set(key, value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{any})\cr
Key.}
      \item{\code{value}}{(\code{any})\cr
Value}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-has"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-has}{}}}
\subsection{\code{LRUCache$has()}}{
  Check whether a given \code{key} exists in the cache.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$has(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{any})\cr
Key.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-remove"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-remove}{}}}
\subsection{\code{LRUCache$remove()}}{
  Remove \code{key} from the cache, returning the value if it was present or \code{NULL} otherwise.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$remove(key)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{any})\cr
Key.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-clear"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-clear}{}}}
\subsection{\code{LRUCache$clear()}}{
  Clear all entries from the cache.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-keys_mru_to_lru"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-keys_mru_to_lru}{}}}
\subsection{\code{LRUCache$keys_mru_to_lru()}}{
  Return keys ordered from most-recently-used to least-recently-used.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$keys_mru_to_lru()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LRUCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LRUCache-clone}{}}}
\subsection{\code{LRUCache$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LRUCache$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
