% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq0.R
\name{seq0}
\alias{seq0}
\alias{seq_len0}
\alias{seq_along0}
\title{Sequence of length 0}
\usage{
seq_len0(n)

seq_along0(x)
}
\arguments{
\item{n}{(\code{integer(1)})\cr
The length of the sequence.}

\item{x}{(\code{integer()})\cr
The vector to sequence along.}
}
\value{
(\code{integer()})
}
\description{
Like \code{\link{seq_len}} and \code{\link{seq_along}} but starts at 0.
}
