% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapevec_repr.R
\name{shapevec_repr}
\alias{shapevec_repr}
\title{Format shape vector as string}
\usage{
shapevec_repr(shape)
}
\arguments{
\item{shape}{(\code{integer()})\cr
A shape vector.}
}
\value{
(\code{character(1)})\cr
The formatted shape string.
}
\description{
Formats a shape vector as a string like \verb{(2,3,4)}.
NA values are replaced with \verb{?}.
}
