% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapevec_repr.R
\name{shapevec_reprs}
\alias{shapevec_reprs}
\title{Format multiple shape vectors}
\usage{
shapevec_reprs(...)
}
\arguments{
\item{...}{\cr
Shape vectors to format.}
}
\value{
(\code{character(1)})\cr
The formatted shapes, separated by ", ".
}
\description{
Applies \code{\link[=shapevec_repr]{shapevec_repr()}} to multiple shapes and combines them.
}
