% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without.R
\name{without}
\alias{without}
\title{Remove elements from a vector}
\usage{
without(x, indices)
}
\arguments{
\item{x}{(\code{vector})\cr
The vector to remove elements from.}

\item{indices}{(\code{integer()})\cr
The indices to remove.}
}
\value{
(\code{vector})\cr
The vector with the elements removed.
}
\description{
Remove elements from a vector by index.
Also works for empty indices.
}
