% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlamisc-package.R
\docType{package}
\name{xlamisc-package}
\alias{xlamisc}
\alias{xlamisc-package}
\title{xlamisc: Helper Functions for 'r-xla'}
\description{
Frequently used generics and helper functions used in 'anvl' and its companion packages 'stablehlo' and 'pjrt'. Comes with S3 generics to query tensors (shape, data type, device) and to convert them to R arrays and raw vectors, as well as an enum-style class for tensor element types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-xla.github.io/xlamisc/}
  \item \url{https://github.com/r-xla/xlamisc}
  \item Report bugs at \url{https://github.com/r-xla/xlamisc/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})

Authors:
\itemize{
  \item Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})
  \item Daniel Falbel \email{daniel@posit.co} (\href{https://orcid.org/0009-0006-0143-2392}{ORCID})
}

}
\keyword{internal}
