% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of.R
\name{new_list_of}
\alias{new_list_of}
\title{Create a new list of class}
\usage{
new_list_of(class_name, item_class, validator = NULL)
}
\arguments{
\item{class_name}{The name of the class}

\item{item_class}{The class of the items in the list}

\item{validator}{A validator function}
}
\value{
A list of class
}
\description{
Create a new list of class
}
