% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlamisc-package.R
\docType{package}
\name{xlamisc-package}
\alias{xlamisc}
\alias{xlamisc-package}
\title{xlamisc: Helper Functions for the XLA Ecosystem}
\description{
Helper functions for the XLA ecosystem.
}
\author{
\strong{Maintainer}: Sebastian Fischer \email{seb.fischer@tutamail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})

Authors:
\itemize{
  \item Daniel Falbel \email{daniel@posit.co} (\href{https://orcid.org/0009-0006-0143-2392}{ORCID})
}

}
